extract some layer related code into: `src/app/core/layers/mixins`
