+++ src/app/core/layers/layer.js
@@ -1,10 +1,9 @@
 import ApplicationState     from 'store/application-state';
-import { DOWNLOAD_FORMATS } from 'app/constant';
 import DataRouterService    from 'services/data';
 import ProjectsRegistry     from 'store/projects';
-import ApplicationService   from 'services/application';
-import GUI                  from 'services/gui';
-import { prompt }           from 'utils/prompt';
+import DownloadMixin   from 'app/core/layers/mixins/download';
+import SelectionMixin  from 'app/core/layers/mixins/selection';
+import RelationsMixin  from 'app/core/layers/mixins/relations';
 
 const { t }                 = require('core/i18n/i18n.service');
 const {
@@ -17,7 +16,6 @@ const {
   geometryFields,
   parseAttributes,
 }                           =  require('utils/geo');
-const Relations             = require('core/relations/relations');
 const ProviderFactory       = require('core/layers/providersfactory');
 
 const deprecate             = require('util-deprecate');
@@ -226,6 +224,10 @@ function Layer(config={}, options={}) {
 
 inherit(Layer, G3WObject);
 
+Object.assign(Layer.prototype, DownloadMixin);
+Object.assign(Layer.prototype, RelationsMixin);
+Object.assign(Layer.prototype, SelectionMixin);
+
 const proto = Layer.prototype;
 
 /**
@@ -328,147 +330,7 @@ proto.getSearchEndPoint = function() {
   return this.getType() !== Layer.LayerTypes.TABLE ? this.config.search_endpoint : 'api';
 };
 
-/**
- * [LAYER RELATIONS]
- * 
- * Create Relation
- * 
- * @param projectRelations
- * 
- * @returns {Relations}
- * 
- * @private
- */
-proto._createRelations = function(projectRelations) {
-  const layerId = this.getId();
-  return new Relations({
-    relations: projectRelations.filter(relation => -1 !== [relation.referencedLayer, relation.referencingLayer].indexOf(layerId))
-  });
-};
 
-/**
- * [LAYER RELATIONS]
- * 
- * Get Relations
- * 
- * @returns {*}
- */
-proto.getRelations = function() {
-  return this._relations
-};
-
-/**
- * [LAYER RELATIONS]
- * 
- * Get Relation by id
- * 
- * @param id
- * 
- * @returns {*}
- */
-proto.getRelationById = function(id) {
-  return this._relations.getArray().find(relation => relation.getId() === id);
-};
-
-/**
- * [LAYER RELATIONS]
- * 
- * Get Relation fields
- * 
- * @param relationName
- * 
- * @returns { * | Array }
- */
-proto.getRelationAttributes = function(relationName) {
-  const relation = this._relations.find(relation => relation.name === relationName);
-  return relation ? relation.fields : [];
-};
-
-/**
- * [LAYER RELATIONS]
- * 
- * @TODO Add description
- * 
- * @returns { Object }
- */
-proto.getRelationsAttributes = function() {
-  const fields = {};
-  this.state.relations.forEach(relation => fields[relation.name] = relation.fields);
-  return fields;
-};
-
-/**
- * [LAYER RELATIONS]
- * 
- * Check if layer is a Child of a relation
- * 
- * @returns { * | boolean }
- */
-proto.isChild = function() {
-  return this.getRelations() ? this._relations.isChild(this.getId()) : false;
-};
-
-/**
- * [LAYER RELATIONS]
- * 
- * Check if layer is a Father of a relation
- * 
- * @returns { * | boolean }
- */
-proto.isFather = function() {
-  return this.getRelations() ? this._relations.isFather(this.getId()) : false;
-};
-
-/**
- * [LAYER RELATIONS]
- * 
- * Get children relations
- * 
- * @returns { * |Array }
- */
-proto.getChildren = function() {
-  return this.isFather() ? this._relations.getChildren(this.getId()) : [];
-};
-
-/**
- * [LAYER RELATIONS]
- * 
- * Get parents relations
- * 
- * @returns { * | Array }
- */
-proto.getFathers = function() {
-  return this.isChild() ? this._relations.getFathers(this.getId()) : [];
-};
-
-/**
- * [LAYER RELATIONS]
- * 
- * Check if it has children
- * 
- * @returns { * | boolean }
- */
-proto.hasChildren = function() {
-  return this.hasRelations() ? this._relations.hasChildren(this.getId()) : false;
-};
-
-/**
- * [LAYER RELATIONS]
- * 
- * Check if it has fathers
- * 
- * @returns { * | boolean }
- */
-proto.hasFathers = function() {
-  return this.hasRelations() ? this._relations.hasFathers(this.getId()) : false;
-};
-
-/**
- * [LAYER RELATIONS]
- */
-proto.hasRelations = function() {
-  return !!this._relations;
-};
 
 /**
  * @TODO Add description
@@ -490,628 +352,7 @@ proto.getAttributeTablePageLength = function() {
 
 // end global state
 
-/**
- * [LAYER SELECTION]
- * 
- * Check if is selected
- * 
- * @returns {boolean}
- */
-proto.isSelected = function() {
-  return this.state.selected;
-};
 
-/**
- * [LAYER SELECTION]
- * 
- * Set Selected
- * 
- * @param {boolean} bool
- */
-proto.setSelected = function(bool) {
-  this.state.selected = bool;
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Set Selection
- * 
- * @param bool
- * 
- * @returns {Promise<void>}
- * 
- * @fires unselectionall
- */
-proto.setSelection = async function(bool=false) {
-  this.state.selection.active = bool;
-
-  // skip when ..
-  if (bool) {
-    return;
-  }
-
-  const is_active   = this.state.filter.active;
-  const has_current = null !== this.state.filter.current;
-
-  /** @TODO add description */
-  if (has_current && is_active) {
-    await this._applyFilterToken(this.state.filter.current.fid)
-  }
-
-  /** @TODO add description */
-  if (!has_current && is_active) {
-    await this.deleteFilterToken();
-  }
-
-  this.emit('unselectionall', this.getId());
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * @returns {boolean} whether selection si active
- */
-proto.isSelectionActive = function() {
-  return this.state.selection.active;
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Get selection
- * 
- * @returns {{ active: boolean }}
- */
-proto.getSelection = function() {
-  return this.state.selection;
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * @returns filter
- */
-proto.getFilter = function() {
-  return this.state.filter;
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Set filter Ative to layer
- * 
- * @param {boolean} bool
- */
-proto.setFilter = function(bool = false) {
-  this.state.filter.active = bool;
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * get current filter
- */
-proto.getCurrentFilter = function() {
-  return this.state.filter.current;
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * @TODO Add description
- * 
- * @returns {boolean}
- */
-proto.getFilterActive = function() {
-  return this.state.filter.active;
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * @returns { Array } saved filters
- */
-proto.getFilters = function() {
-  return this.state.filters;
-}
-
-/**
- * [LAYER SELECTION]
- * 
- * Add new filter
- * 
- * @param filter Object filter
- */
-proto.addFilter = function(filter = {}) {
-  this.state.filters.push(filter);
-}
-
-/**
- * [LAYER SELECTION]
- * 
- * Remove saved filter from filters Array
- * 
- * @param fid unique filter id
- */
-proto.removefilter = function(fid) {
-  this.state.filters = this.state.filters.filter(f => fid === f.fid);
-}
-
-/**
- * [LAYER SELECTION]
- * 
- * Set Current filter
- * 
- * @param {{ fid, name }} filter 
- */
-proto.setCurrentFilter = function(filter) {
-  this.state.filter.current = filter;
-}
-
-/**
- * [LAYER SELECTION]
- * 
- * Apply layer filter by fid
- * 
- * @param filter
- */
-proto.applyFilter = async function(filter) {
-  if (!this.providers['filtertoken']) {
-    return;
-  }
-
-  // current filter is set and is different from current
-  if (null === this.state.filter.current || filter.fid !== this.state.filter.current.fid ) {
-    await this.clearSelectionFids();
-    GUI.closeContent();
-  }
-
-  await this._applyFilterToken(filter);
-}
-
-/**
- * [LAYER SELECTION]
- * 
- * @returns {Promise<void>}
- * 
- * @private
- */
-proto._applyFilterToken = async function(filter) {
-  const { filtertoken } = await this.providers['filtertoken'].applyFilterToken(filter.fid);
-  if (filtertoken) {
-    this.setFilter(false);
-    this.setCurrentFilter(filter);
-    this.setFilterToken(filtertoken);
-  }
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * @since 3.9.0
- */
-proto.saveFilter = function() {
-
-  // skip when ..
-  if (!this.providers['filtertoken'] || !this.selectionFids.size > 0) {
-    return;
-  }
-
-  const layer = this;
-
-  prompt({
-    label: t('layer_selection_filter.tools.savefilter'),
-    value: layer.getCurrentFilter() ? layer.getCurrentFilter().name : '' ,
-    callback: async(name) => {
-      const data = await layer.providers['filtertoken'].saveFilterToken(name);
-
-      // skip when no data return from provider
-      if (!data) {
-        return;
-      }
-    
-      let filter = layer.state.filters.find(f => f.fid === data.fid);
-    
-      // add saved filter to filters array
-      if (undefined === filter) {
-        filter = {
-          fid: data.fid, //get fid
-          name: data.name //get name
-        }
-        layer.state.filters.push(filter);
-      }
-
-      layer.setCurrentFilter(filter);      // set current filter
-      layer.setFilter(false);              // set to false
-      layer.getSelection().active = false; // reset selection to false
-      layer.selectionFids.clear();         // clear current fids
-    
-      //in case of geolayer
-      if (layer.isGeoLayer()) {
-        //remove selection feature from map
-        layer.setOlSelectionFeatures();
-      }
-    
-      //emit unselectionall
-      layer.emit('unselectionall', layer.getId());
-    
-    },
-  });
-
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Method to set unset filter token on layer
- */
-proto.toggleFilterToken = async function() {
-
-  // toggle boolean value of filter active
-  this.setFilter(!this.state.filter.active);
-
-  const has_current = this.state.filter.current;
-  const is_active   = this.state.filter.active;
-
-  // there is an active filter --> create a new filter
-  if (is_active) {
-    await this.createFilterToken();
-  }
-
-  // there is a current saved filter --> apply filter
-  if (has_current && !is_active) {
-    await this.applyFilter(this.state.filter.current);
-  }
-
-  // there is no current saved filter --> delete it
-  if (!has_current && !is_active) {
-    await this.deleteFilterToken();
-  }
-
-  const has_selection = this.state.selection.active && this.isGeoLayer();
-
-  // active filter --> hide all selected feature from map (red ones)
-  if (has_selection && this.state.filter.active) {
-    this.hideOlSelectionFeatures();
-  }
-
-  // active filter --> show only current selected feature on map (red ones)
-  if (has_selection && !this.state.filter.active){
-    this.showAllOlSelectionFeatures();
-  }
-
-  return this.state.filter.active;
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Delete filtertoken from server
- * 
- * @param fid  unique id of filter saved to delete
- */
-proto.deleteFilterToken = async function(fid) {
-  try {
-    // skip when no filtertoken provider is set
-    if (!this.providers['filtertoken']) {
-      return;
-    }
-
-    // delete filtertoken related to layer provider
-    const filtertoken = await this.providers['filtertoken'].deleteFilterToken(fid);
-
-    // remove it from filters list when deleting a saved filter (since v3.9.0)
-    if (undefined !== fid) {
-      this.state.filters = this.state.filters.filter(f => fid !== f.fid);
-    }
-
-    this.setCurrentFilter(null);      // set current filter set to null
-    this.setFilter(false);            // set active filter to false
-    this.setFilterToken(filtertoken); // pass `filtertoken` to application
-
-  } catch(err) {
-    console.log('Error deleteing filtertoken', err);
-  }
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Set applicaton filter token
- * 
- * @param filtertoken
- *
- * @fires filtertokenchange when filtertoken is changed
- * 
- * @since 3.9.0
- */
-proto.setFilterToken = function(filtertoken = null) {
-  ApplicationService.setFilterToken(filtertoken);
-  this.emit('filtertokenchange', { layerId: this.getId() });
-}
-
-/**
- * [LAYER SELECTION]
- *  
- * Create filter token
- */
-proto.createFilterToken = async function() {
-  try {
-
-    const provider  = this.providers['filtertoken'];
-    const selection = this.selectionFids;
-
-    // skip when no filter token provider is set or selectionFids is empty
-    if (!provider || !selection.size > 0) {
-      return;
-    }
-
-    // select all features
-    if (selection.has(Layer.SELECTION_STATE.ALL)) {
-      await provider.deleteFilterToken();
-      this.setFilterToken(null);
-      return;
-    }
-
-    const fids = Array.from(selection);
-
-    // exclude some features from selection
-    if (selection.has(Layer.SELECTION_STATE.EXCLUDE)) {
-      this.setFilterToken( await provider.getFilterToken({ fidsout: fids.filter(id => id !== Layer.SELECTION_STATE.EXCLUDE).join(',') }) );
-      return;
-    }
-
-    // include some features in selection
-    this.setFilterToken( await provider.getFilterToken({ fidsin: fids.join(',') }) );
-
-  } catch(err) {
-    console.log('Error create update token', err);
-  }
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Get Application filter token
- * 
- * @returns {*}
- */
-proto.getFilterToken = function () {
-  return ApplicationService.getFilterToken();
-};
-// end filter token
-
-/**
- * [LAYER SELECTION]
- */
-proto.setSelectionFidsAll = function() {
-  this.selectionFids.clear();
-  this.selectionFids.add(Layer.SELECTION_STATE.ALL);
-
-  /** @TODO add description */
-  if (this.isGeoLayer()) {
-    this.showAllOlSelectionFeatures();
-  }
-
-  /** @TODO add description */
-  this.setSelection(true);
-  if (this.state.filter.active) {
-    this.createFilterToken();
-  }
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * @returns {Set<any>} stored selection `fids` 
- */
-proto.getSelectionFids = function() {
-  return this.selectionFids;
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Invert current selection fids
- */
-proto.invertSelectionFids = function() {
-
-  const selection = this.selectionFids;
-
-  /** @TODO add description */
-  if (selection.has(Layer.SELECTION_STATE.EXCLUDE))  { selection.delete(Layer.SELECTION_STATE.EXCLUDE); }
-  else if (selection.has(Layer.SELECTION_STATE.ALL)) { selection.delete(Layer.SELECTION_STATE.ALL); }
-  else if (selection.size > 0)                       { selection.add(Layer.SELECTION_STATE.EXCLUDE); }
-
-  /** @TODO add description */
-  if (this.isGeoLayer()) {
-    this.setInversionOlSelectionFeatures();
-  }
-
-  /** @TODO add description */
-  if (this.state.filter.active) {
-    this.createFilterToken();
-  }
-
-  this.setSelection(selection.size > 0);
-};
-
-
-/**
- * [LAYER SELECTION]
- * 
- * Check if feature id is present
- * 
- * @param fid feature id
- * 
- * @returns {boolean}
- */
-proto.hasSelectionFid = function(fid) {
-  /** @TODO add description */
-  if (this.selectionFids.has(Layer.SELECTION_STATE.ALL)) {
-    return true;
-  }
-  /** @TODO add description */
-  if (this.selectionFids.has(Layer.SELECTION_STATE.EXCLUDE)) {
-    return !this.selectionFids.has(fid);
-  }
-  /** @TODO add description */
-  return this.selectionFids.has(fid);
-};
-
-
-/**
- * [LAYER SELECTION]
- * 
- * Include fid feature id to selection
- * 
- * @param fid
- * @param createToken
- * 
- * @returns {Promise<void>}
- */
-proto.includeSelectionFid = async function(fid, createToken=true) {
-  //set create filter token
-  //check if fid is excluded from selection
-  const excludeFidFromSelection = (
-    this.selectionFids.has(Layer.SELECTION_STATE.EXCLUDE) &&
-    this.selectionFids.has(fid)
-  )
-
-  // remove fid
-  if (excludeFidFromSelection) {
-    this.selectionFids.delete(fid);
-  }
-
-  // if the only one exclude set all selected
-  if (excludeFidFromSelection && 1 === this.selectionFids.size) {
-    this.setSelectionFidsAll();
-  }
-
-  // add to selction fid
-  if (!excludeFidFromSelection) {
-    this.selectionFids.add(fid);
-  }
-
-  /** @TODO add description */
-  if (!excludeFidFromSelection && !this.isSelectionActive()) {
-    this.setSelection(true);
-  }
-  
-  /** @TODO add description */
-  if (this.isGeoLayer()) {
-   this.setOlSelectionFeatureByFid(fid, 'add');
-  }
-  
-  /** @TODO add description */
-  if (createToken && this.state.filter.active) {
-    await this.createFilterToken();
-  }
-
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Exclude fid to selection
- * 
- * @param fid
- * @param createToken
- * 
- * @returns {Promise<void>}
- */
-proto.excludeSelectionFid = async function(fid, createToken = true) {
-
-  const selection = this.selectionFids;
-
-  /** @TODO add description */
-  if (selection.has(Layer.SELECTION_STATE.ALL) || 0 === selection.size) {
-    selection.clear();
-    selection.add(Layer.SELECTION_STATE.EXCLUDE);
-  }
-
-  /** @TODO add description */
-  if (selection.has(Layer.SELECTION_STATE.EXCLUDE)) {
-    selection.add(fid);
-  } else {
-    selection.delete(fid);
-  }
-
-  /** @TODO add description */
-  if (1 === selection.size && selection.has(Layer.SELECTION_STATE.EXCLUDE)) {
-    this.setselectionFidsAll();
-  }
-
-  const isLastFeatureSelected  = this.isGeoLayer() && this.setOlSelectionFeatureByFid(fid, 'remove');
-
-  /** @TODO add description */
-  if (createToken && this.state.filter.active) {
-    await this.createFilterToken();
-  }
-
-  /** @TODO add description */
-  if (0 === selection.size || isLastFeatureSelected) {
-    selection.clear();
-    this.setSelection(false);
-  }
-
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * @param { Array }   fids
- * @param { boolean } createToken since 3.9.0
- * 
- * @returns { Promise<void> }
- */
-proto.includeSelectionFids = async function(fids = [], createToken = true) {
-  // pass false because eventually token filter creation need to be called after
-  fids.forEach(fid => this.includeSelectionFid(fid, false));
-
-  /** @TODO add description */
-  if (createToken && this.state.filter.active) {
-    await this.createFilterToken();
-  }
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Exclude fids from selection
- * 
- * @param { Array }   fids
- * @param { boolean } createToken since 3.9.0
- * 
- * @returns { Promise<void> }
- */
-proto.excludeSelectionFids = async function(fids = [], createToken = true) {
-  //pass false because eventually token filter creation need to be called after
-  fids.forEach(fid => this.excludeSelectionFid(fid, false));
-
-  /** @TODO add description */
-  if (createToken && this.state.filter.active) {
-    await this.createFilterToken();
-  }
-};
-
-/**
- * [LAYER SELECTION]
- * 
- * Clear selection
- */
-proto.clearSelectionFids = async function() {
-  //cLear set selection fids
-  this.selectionFids.clear();
-  //if ia a layer with geometry
-  if (this.isGeoLayer()) {
-    //remove feature selected on map
-    this.setOlSelectionFeatures();
-  }
-  //set selection false
-  await this.setSelection(false);
-};
 
 /**
  * Return wms layer name for wms request
@@ -1129,110 +370,6 @@ proto.isWmsUseLayerIds = function() {
   return this.config.wms_use_layer_ids;
 };
 
-
-/**
- *
- * DOWNLOAD METHODS
- */
-
-/** 
- * @returns promise
- */
-proto.getDownloadFilefromDownloadDataType = function(type, {data={}, options}) {
-  data.filtertoken = this.getFilterToken();
-  switch (type) {
-    case 'shapefile': return this.getShp({data, options});
-    case 'xls':       return this.getXls({data, options});
-    case 'csv':       return this.getCsv({data, options});
-    case 'gpx':       return this.getGpx({data, options});
-    case 'gpkg':      return this.getGpkg({data, options});
-    case 'geotiff':   return this.getGeoTIFF({ data, options });
-  }
-};
-
-/**
- * Get Getotiff layer format
- * @param data
- * @returns {Promise | Promise<unknown>}
- */
-proto.getGeoTIFF = function({data={}}={}) {
-  data.filtertoken = this.getFilterToken();
-  return XHR.fileDownload({
-    url: this.getUrl('geotiff'),
-    data,
-    httpMethod: "POST"
-  })
-};
-
-/**
- * Get Xls layer format
- * @param data
- * @returns {Promise | Promise<unknown>}
- */
-proto.getXls = function({data={}}={}) {
-  data.filtertoken = this.getFilterToken();
-  return XHR.fileDownload({
-    url: this.getUrl('xls'),
-    data,
-    httpMethod: "POST"
-  })
-};
-
-/**
- * Get shapefile layer format
- * @param data
- * @returns {Promise | Promise<unknown>}
- */
-proto.getShp = function({data={}}={}) {
-  data.filtertoken = this.getFilterToken();
-  return XHR.fileDownload({
-    url: this.getUrl('shp'),
-    data,
-    httpMethod: "POST"
-  })
-};
-
-/**
- * Get gpx layer format
- * @param data
- * @returns {Promise | Promise<unknown>}
- */
-proto.getGpx = function({data={}}={}) {
-  data.filtertoken = this.getFilterToken();
-  return XHR.fileDownload({
-    url: this.getUrl('gpx'),
-    data,
-    httpMethod: "POST"
-  })
-};
-
-/**
- * get gpkg layer format
- * @param data
- * @returns {Promise | Promise<unknown>}
- */
-proto.getGpkg = function({data={}}={}) {
-  data.filtertoken = this.getFilterToken();
-  return XHR.fileDownload({
-    url: this.getUrl('gpkg'),
-    data,
-    httpMethod: "POST"
-  })
-};
-
-/**
- * Get csv layer format
- * @param data
- * @returns {Promise | Promise<unknown>}
- */
-proto.getCsv = function({data={}}={}) {
-  data.filtertoken = this.getFilterToken();
-  return XHR.fileDownload({
-    url: this.getUrl('csv'),
-    data,
-    httpMethod: "POST"
-  })
-};
 
 /**
  * Get source type of layer
@@ -1573,89 +710,6 @@ proto.getState = function() {
  */
 proto.getSource = function() {
   return this.state.source;
-};
-
-/**
- * Check if it has a format to download
- * @returns {*}
- */
-proto.isDownloadable = function() {
-  return (
-    this.isShpDownlodable()  ||
-    this.isXlsDownlodable()  ||
-    this.isGpxDownlodable()  ||
-    this.isGpkgDownlodable() ||
-    this.isCsvDownlodable()
-  );
-};
-
-/**
- * Get downlaod formats
- * @returns {string[]}
- */
-proto.getDownloadableFormats = function() {
-  return Object
-    .keys(DOWNLOAD_FORMATS)
-    .filter(download_format => this.config[download_format])
-    .map(format => DOWNLOAD_FORMATS[format].format);
-};
-
-/**
- *
- * @param download url
- * @returns {string}
- */
-proto.getDownloadUrl = function(format) {
-  const find = Object
-    .values(DOWNLOAD_FORMATS)
-    .find(download_format => download_format.format === format);
-  return find && find.url;
-};
-
-/**
- *
- * @returns {false|*|boolean}
- */
-proto.isGeoTIFFDownlodable = function() {
-  return !this.isBaseLayer() && this.config.download && 'gdal' === this.config.source.type ;
-};
-
-/**
- *
- * @returns {false|*|boolean}
- */
-proto.isShpDownlodable = function() {
-  return !this.isBaseLayer() && this.config.download && 'gdal' !== this.config.source.type;
-};
-/**
- *
- * @returns {false|string|*}
- */
-proto.isXlsDownlodable = function() {
-  return !this.isBaseLayer() && this.config.download_xls;
-};
-/**
- *
- * @returns {false|string|*}
- */
-proto.isGpxDownlodable = function() {
-  return !this.isBaseLayer() && this.config.download_gpx;
-};
-
-/**
- *
- * @returns {false|string|*}
- */
-proto.isGpkgDownlodable = function() {
-  return !this.isBaseLayer() && this.config.download_gpkg;
-};
-
-/**
- *
- * @returns {false|string|*}
- */
-proto.isCsvDownlodable = function() {
-  return !this.isBaseLayer() && this.config.download_csv;
 };
 
 /**
@@ -2273,14 +1327,6 @@ Layer.EDITOPS = {
   INSERT: 1,
   UPDATE: 2,
   DELETE: 4
-};
-
-/**
- * Selection state 
- */
-Layer.SELECTION_STATE = {
-  ALL: '__ALL__',
-  EXCLUDE: '__EXCLUDE__'
 };
 
 module.exports = Layer;
